requirements.txt
pyserial==3.5
flask
python-dotenv